requirements.txt
Django==5.0.6
djangorestframework==3.15.1
pandas==2.2.3
python-dotenv==1.1.0
pytest==8.0.2
pytest-django==4.8.0 